record exposures through gate eval
I need to dig into the order of exposures to make sure the primary
exposure is listed first and secondary exposures are after (same with
the rule ID that is populated at the top level of the result). Need to
make sure that we're logging the right things as secondary vs. primary.